+++ vck-openid-ktor/src/commonTest/kotlin/at/asitplus/wallet/lib/ktor/openid/OpenId4VciClientTest.kt
@@ -1,18 +1,27 @@
 package at.asitplus.wallet.lib.ktor.openid
 
-import at.asitplus.KmmResult
 import at.asitplus.catching
 import at.asitplus.openid.*
+import at.asitplus.openid.AuthenticationRequestParameters
+import at.asitplus.openid.CredentialRequestParameters
+import at.asitplus.openid.OpenIdConstants
+import at.asitplus.openid.TokenRequestParameters
 import at.asitplus.signum.indispensable.CryptoPublicKey
-import at.asitplus.signum.indispensable.josef.JsonWebToken
 import at.asitplus.signum.indispensable.josef.toJwsAlgorithm
 import at.asitplus.wallet.eupid.EuPidScheme
 import at.asitplus.wallet.lib.agent.*
+import at.asitplus.wallet.lib.agent.ClaimToBeIssued
+import at.asitplus.wallet.lib.agent.CredentialToBeIssued.Iso
+import at.asitplus.wallet.lib.agent.CredentialToBeIssued.VcSd
+import at.asitplus.wallet.lib.agent.EphemeralKeyWithSelfSignedCert
+import at.asitplus.wallet.lib.agent.EphemeralKeyWithoutCert
+import at.asitplus.wallet.lib.agent.IssuerAgent
 import at.asitplus.wallet.lib.agent.Verifier.VerifyCredentialResult
 import at.asitplus.wallet.lib.data.*
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.ISO_MDOC
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.PLAIN_JWT
 import at.asitplus.wallet.lib.data.ConstantIndex.CredentialRepresentation.SD_JWT
+import at.asitplus.wallet.lib.data.vckJsonSerializer
 import at.asitplus.wallet.lib.iso.IssuerSignedItem
 import at.asitplus.wallet.lib.jws.JwsHeaderCertOrJwk
 import at.asitplus.wallet.lib.jws.JwsHeaderJwk
@@ -23,40 +32,56 @@ import at.asitplus.wallet.lib.oauth2.ClientAuthenticationService
 import at.asitplus.wallet.lib.oauth2.SimpleAuthorizationService
 import at.asitplus.wallet.lib.oauth2.RequestInfo
 import at.asitplus.wallet.lib.oauth2.TokenService
-import at.asitplus.wallet.lib.oidvci.*
+import at.asitplus.wallet.lib.oidvci.BuildClientAttestationJwt
+import at.asitplus.wallet.lib.oidvci.CredentialAuthorizationServiceStrategy
+import at.asitplus.wallet.lib.oidvci.CredentialIssuer
+import at.asitplus.wallet.lib.oidvci.CredentialIssuerDataProvider
+import at.asitplus.wallet.lib.oidvci.DefaultNonceService
+import at.asitplus.wallet.lib.oidvci.OAuth2DataProvider
+import at.asitplus.wallet.lib.oidvci.WalletService
+import at.asitplus.wallet.lib.oidvci.decodeFromPostBody
+import at.asitplus.wallet.lib.oidvci.decodeFromUrlQuery
 import com.benasher44.uuid.uuid4
 import io.github.aakira.napier.Napier
 import io.kotest.core.spec.style.FunSpec
 import io.kotest.matchers.booleans.shouldBeTrue
+import io.kotest.matchers.collections.shouldBeSingleton
 import io.kotest.matchers.nulls.shouldNotBeNull
 import io.kotest.matchers.shouldBe
 import io.kotest.matchers.types.shouldBeInstanceOf
-import io.ktor.client.*
-import io.ktor.client.engine.*
-import io.ktor.client.engine.mock.*
+import io.ktor.client.HttpClient
+import io.ktor.client.engine.mock.MockEngine
+import io.ktor.client.engine.mock.respond
+import io.ktor.client.engine.mock.respondError
+import io.ktor.client.engine.mock.respondRedirect
+import io.ktor.client.engine.mock.toByteArray
 import io.ktor.client.request.*
-import io.ktor.http.*
-import io.ktor.util.*
-import kotlinx.coroutines.Dispatchers
-import kotlinx.coroutines.sync.Mutex
+import io.ktor.client.request.get
+import io.ktor.http.ContentType
+import io.ktor.http.HttpHeaders
+import io.ktor.http.HttpStatusCode
+import io.ktor.http.fullPath
+import io.ktor.http.headersOf
+import io.ktor.util.toMap
 import kotlinx.coroutines.test.runTest
-import kotlinx.coroutines.withContext
-import kotlinx.coroutines.withTimeout
 import kotlinx.datetime.Clock
 import kotlinx.serialization.json.jsonPrimitive
 import kotlin.random.Random
 
 class OpenId4VciClientTest : FunSpec() {
 
-    lateinit var countdownLatch: Mutex
     lateinit var credentialKeyMaterial: KeyMaterial
     lateinit var dpopKeyMaterial: KeyMaterial
     lateinit var clientAuthKeyMaterial: KeyMaterial
     lateinit var refreshTokenStore: RefreshTokenInfo
 
+    lateinit var mockEngine: MockEngine
+    lateinit var credentialIssuer: CredentialIssuer
+    lateinit var authorizationService: SimpleAuthorizationService
+    lateinit var client: OpenId4VciClient
+
     init {
         beforeEach {
-            countdownLatch = Mutex(true)
             credentialKeyMaterial = EphemeralKeyWithoutCert()
             dpopKeyMaterial = EphemeralKeyWithoutCert()
             clientAuthKeyMaterial = EphemeralKeyWithoutCert()
@@ -65,22 +90,10 @@ class OpenId4VciClientTest : FunSpec() {
         test("loadEuPidCredentialSdJwt") {
             runTest {
                 val expectedFamilyName = uuid4().toString()
-                val (client, credentialIssuer) = setup(
+                setup(
                     scheme = EuPidScheme,
                     representation = SD_JWT,
-                    attributes = mapOf(
-                        EuPidScheme.Attributes.FAMILY_NAME to expectedFamilyName
-                    ),
-                    storeCredential = {
-                        it.shouldBeInstanceOf<Holder.StoreCredentialInput.SdJwt>()
-                        it.scheme shouldBe EuPidScheme
-                        val sdJwt =
-                            Validator().verifySdJwt(SdJwtSigned.parse(it.vcSdJwt)!!, credentialKeyMaterial.publicKey)
-                        sdJwt.shouldBeInstanceOf<VerifyCredentialResult.SuccessSdJwt>()
-                        sdJwt.disclosures.values.any { it.claimName == EuPidScheme.Attributes.FAMILY_NAME && it.claimValue.jsonPrimitive.content == expectedFamilyName }
-                            .shouldBeTrue()
-                        countdownLatch.unlock()
-                    }
+                    attributes = mapOf(EuPidScheme.Attributes.FAMILY_NAME to expectedFamilyName),
                 )
 
                 // Load credential identifier infos from Issuing service
@@ -89,40 +102,35 @@ class OpenId4VciClientTest : FunSpec() {
                 val selectedCredential = credentialIdentifierInfos
                     .first { it.supportedCredentialFormat.format == CredentialFormatEnum.DC_SD_JWT }
                 // client will call clientBrowser.openUrlExternally
-                client.startProvisioningWithAuthRequest(
+                client.startProvisioningWithAuthRequestReturningResult(
                     credentialIssuerUrl = "http://localhost",
                     credentialIdentifierInfo = selectedCredential,
-                ).apply {
-                    this.isSuccess shouldBe true
+                ).getOrThrow().also {
+                    // Simulates the browser, handling authorization to get the authCode
+                    val httpClient = HttpClient(mockEngine) { followRedirects = false }
+                    val authCode = httpClient.get(it.url).headers[HttpHeaders.Location]
+                    client.resumeWithAuthCode(authCode!!, it.context).getOrThrow().also {
+                        refreshTokenStore = it.refreshToken!!
+                        verifySdJwtCredential(it, expectedFamilyName)
+                    }
                 }
-                assertCorrectCredentialIssued()
 
-                countdownLatch = Mutex(true)
                 refreshTokenStore.shouldNotBeNull()
-                client.refreshCredential(refreshTokenStore).apply {
-                    this.isSuccess shouldBe true
+                client.refreshCredentialReturningResult(refreshTokenStore).getOrThrow().also {
+                    verifySdJwtCredential(it, expectedFamilyName)
                 }
-                assertCorrectCredentialIssued()
             }
         }
 
         test("loadEuPidCredentialIsoWithOffer") {
             runTest {
                 val expectedGivenName = uuid4().toString()
-                val (client, credentialIssuer, authorizationService) = setup(
+                setup(
                     scheme = EuPidScheme,
                     representation = ISO_MDOC,
                     attributes = mapOf(
                         EuPidScheme.Attributes.GIVEN_NAME to expectedGivenName
-                    ),
-                    storeCredential = {
-                        it.shouldBeInstanceOf<Holder.StoreCredentialInput.Iso>()
-                        it.scheme shouldBe EuPidScheme
-                        it.issuerSigned.namespaces?.values?.flatMap { it.entries }?.map { it.value }
-                            ?.any { it.elementIdentifier == EuPidScheme.Attributes.GIVEN_NAME && it.elementValue == expectedGivenName }
-                            ?.shouldNotBeNull()?.shouldBeTrue()
-                        countdownLatch.unlock()
-                    }
+                    )
                 )
 
                 // Load credential identifier infos from Issuing service
@@ -135,17 +143,49 @@ class OpenId4VciClientTest : FunSpec() {
                     dummyUser(),
                     credentialIssuer.metadata.credentialIssuer
                 )
-                client.loadCredentialWithOffer(offer, selectedCredential, null).apply {
-                    this.isSuccess shouldBe true
+                client.loadCredentialWithOfferReturningResult(offer, selectedCredential, null).getOrThrow().also {
+                    it.shouldBeInstanceOf<CredentialIssuanceResult.Success>().also {
+                        refreshTokenStore = it.refreshToken!!
+                        verifyIsoMdocCredential(it, expectedGivenName)
+                    }
                 }
-                assertCorrectCredentialIssued()
-
-                countdownLatch = Mutex(true)
                 refreshTokenStore.shouldNotBeNull()
-                client.refreshCredential(refreshTokenStore).apply {
-                    this.isSuccess shouldBe true
+                client.refreshCredentialReturningResult(refreshTokenStore).getOrThrow().also {
+                    verifyIsoMdocCredential(it, expectedGivenName)
                 }
-                assertCorrectCredentialIssued()
+            }
+        }
+    }
+
+    private suspend fun verifySdJwtCredential(
+        success: CredentialIssuanceResult.Success,
+        expectedFamilyName: String,
+    ) {
+        success.credentials.shouldBeSingleton().also {
+            it.first().shouldBeInstanceOf<Holder.StoreCredentialInput.SdJwt>().also {
+                it.scheme shouldBe EuPidScheme
+                val sdJwt =
+                    Validator().verifySdJwt(
+                        SdJwtSigned.parse(it.vcSdJwt)!!,
+                        credentialKeyMaterial.publicKey
+                    )
+                sdJwt.shouldBeInstanceOf<VerifyCredentialResult.SuccessSdJwt>()
+                sdJwt.disclosures.values.any { it.claimName == EuPidScheme.Attributes.FAMILY_NAME && it.claimValue.jsonPrimitive.content == expectedFamilyName }
+                    .shouldBeTrue()
+            }
+        }
+    }
+
+    private fun verifyIsoMdocCredential(
+        success: CredentialIssuanceResult.Success,
+        expectedGivenName: String,
+    ) {
+        success.credentials.shouldBeSingleton().also {
+            it.first().shouldBeInstanceOf<Holder.StoreCredentialInput.Iso>().also {
+                it.scheme shouldBe EuPidScheme
+                it.issuerSigned.namespaces?.values?.flatMap { it.entries }?.map { it.value }
+                    ?.any { it.elementIdentifier == EuPidScheme.Attributes.GIVEN_NAME && it.elementValue == expectedGivenName }
+                    ?.shouldNotBeNull()?.shouldBeTrue()
             }
         }
     }
@@ -154,83 +194,25 @@ class OpenId4VciClientTest : FunSpec() {
         scheme: ConstantIndex.CredentialScheme,
         representation: ConstantIndex.CredentialRepresentation,
         attributes: Map<String, String>,
-        storeCredential: (suspend (Holder.StoreCredentialInput) -> Unit) = {},
-    ): SetupResult {
-        val (mockEngine, credentialIssuer, authorizationService) = setupIssuingService(
-            scheme,
-            representation,
-            attributes
-        )
-        val client = setupClient(mockEngine, storeCredential)
-        return SetupResult(client, credentialIssuer, authorizationService)
-    }
-
-    private fun setupClient(
-        mockEngine: HttpClientEngine,
-        storeCredential: (suspend (Holder.StoreCredentialInput) -> Unit),
-    ): OpenId4VciClient {
-        // This construction is needed to continue with client in the openUrlExternally callback
-        var client: OpenId4VciClient? = null
-        // Simulates the browser, handling authorization to get the authCode
-        val clientBrowser = object {
-            suspend fun openUrlExternally(it: String) {
-                val httpClient = HttpClient(mockEngine) { followRedirects = false }
-                val authCode = httpClient.get(it).headers[HttpHeaders.Location]
-                client!!.resumeWithAuthCode(authCode!!)
-            }
-        }
-        var provisioningContextStore: ProvisioningContext? = null
-        val clientId = "https://example.com/rp"
-        client = OpenId4VciClient(
-            openUrlExternally = { clientBrowser.openUrlExternally(it) },
-            engine = mockEngine,
-            storeProvisioningContext = { provisioningContextStore = it },
-            loadProvisioningContext = { provisioningContextStore },
-            loadClientAttestationJwt = {
-                BuildClientAttestationJwt(
-                    SignJwt<JsonWebToken>(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
-                    clientId = clientId,
-                    issuer = "issuer",
-                    clientKey = clientAuthKeyMaterial.jsonWebKey
-                ).serialize()
-            },
-            signClientAttestationPop = SignJwt(clientAuthKeyMaterial, JwsHeaderNone()),
-            signDpop = SignJwt(dpopKeyMaterial, JwsHeaderJwk()),
-            dpopAlgorithm = dpopKeyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
-            oid4vciService = WalletService(
-                clientId = clientId,
-                keyMaterial = credentialKeyMaterial,
-            ),
-            storeCredential = storeCredential,
-            storeRefreshToken = { refreshTokenStore = it }
-        )
-        return client
-    }
-
-    // Mocks the Issuing Service, which will be called by [OpenId4VciClient]
-    private fun setupIssuingService(
-        scheme: ConstantIndex.CredentialScheme,
-        representationToIssue: ConstantIndex.CredentialRepresentation,
-        attributesToIssue: Map<String, String>,
-    ): Triple<HttpClientEngine, CredentialIssuer, SimpleAuthorizationService> {
+    ) {
         val dataProvider = OAuth2DataProvider { _, _ -> dummyUser() }
         val credentialProvider =
             CredentialIssuerDataProvider { _, subjectPublicKey: CryptoPublicKey, credentialScheme, representation, _ ->
                 catching {
                     require(credentialScheme == scheme)
-                    require(representation == representationToIssue)
+                    require(representation == representation)
                     var digestId = 0u
                     when (representation) {
                         PLAIN_JWT -> TODO()
-                        SD_JWT -> CredentialToBeIssued.VcSd(
-                            attributesToIssue.map { ClaimToBeIssued(it.key, it.value) },
+                        SD_JWT -> VcSd(
+                            attributes.map { ClaimToBeIssued(it.key, it.value) },
                             Clock.System.now(),
                             credentialScheme,
                             subjectPublicKey
                         )
 
-                        ISO_MDOC -> CredentialToBeIssued.Iso(
-                            attributesToIssue.map {
+                        ISO_MDOC -> Iso(
+                            attributes.map {
                                 IssuerSignedItem(
                                     digestId++,
                                     Random.nextBytes(32),
@@ -252,7 +234,7 @@ class OpenId4VciClientTest : FunSpec() {
         val nonceEndpointPath = "/nonce"
         val parEndpointPath = "/par"
         val publicContext = "https://issuer.example.com"
-        val authorizationService = SimpleAuthorizationService(
+        authorizationService = SimpleAuthorizationService(
             strategy = CredentialAuthorizationServiceStrategy(credentialSchemes),
             dataProvider = dataProvider,
             publicContext = publicContext,
@@ -268,7 +250,7 @@ class OpenId4VciClientTest : FunSpec() {
                 issueRefreshTokens = true
             ),
         )
-        val credentialIssuer = CredentialIssuer(
+        credentialIssuer = CredentialIssuer(
             authorizationService = authorizationService,
             issuer = IssuerAgent(EphemeralKeyWithSelfSignedCert()),
             credentialSchemes = credentialSchemes,
@@ -277,34 +259,34 @@ class OpenId4VciClientTest : FunSpec() {
             credentialEndpointPath = credentialEndpointPath,
             nonceEndpointPath = nonceEndpointPath,
         )
-
-        return Triple(MockEngine { request ->
+        mockEngine = MockEngine { request ->
             when {
                 request.url.fullPath == OpenIdConstants.PATH_WELL_KNOWN_CREDENTIAL_ISSUER -> respond(
-                    vckJsonSerializer.encodeToString(credentialIssuer.metadata),
+                    vckJsonSerializer.encodeToString<IssuerMetadata>(credentialIssuer.metadata),
                     headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                 )
 
                 request.url.fullPath == OpenIdConstants.PATH_WELL_KNOWN_OPENID_CONFIGURATION -> respond(
-                    vckJsonSerializer.encodeToString(authorizationService.metadata),
+                    vckJsonSerializer.encodeToString<OAuth2AuthorizationServerMetadata>(authorizationService.metadata),
                     headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                 )
 
                 request.url.fullPath == OpenIdConstants.PATH_WELL_KNOWN_OAUTH_AUTHORIZATION_SERVER -> respond(
-                    vckJsonSerializer.encodeToString(authorizationService.metadata),
+                    vckJsonSerializer.encodeToString<OAuth2AuthorizationServerMetadata>(authorizationService.metadata),
                     headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                 )
 
                 request.url.fullPath.startsWith(parEndpointPath) -> {
                     val requestBody = request.body.toByteArray().decodeToString()
-                    val authnRequest: AuthenticationRequestParameters = requestBody.decodeFromPostBody()
+                    val authnRequest: AuthenticationRequestParameters =
+                        requestBody.decodeFromPostBody<AuthenticationRequestParameters>()
                     val result = authorizationService.par(
                         authnRequest,
                         request.headers["OAuth-Client-Attestation"],
                         request.headers["OAuth-Client-Attestation-PoP"]
                     ).getOrThrow()
                     respond(
-                        vckJsonSerializer.encodeToString(result),
+                        vckJsonSerializer.encodeToString<PushedAuthenticationResponseParameters>(result),
                         headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                     )
                 }
@@ -314,18 +296,18 @@ class OpenId4VciClientTest : FunSpec() {
                     val queryParameters: Map<String, String> =
                         request.url.parameters.toMap().entries.associate { it.key to it.value.first() }
                     val authnRequest: AuthenticationRequestParameters =
-                        if (requestBody.isEmpty()) queryParameters.decodeFromUrlQuery()
-                        else requestBody.decodeFromPostBody()
+                        if (requestBody.isEmpty()) queryParameters.decodeFromUrlQuery<AuthenticationRequestParameters>()
+                        else requestBody.decodeFromPostBody<AuthenticationRequestParameters>()
                     val result = authorizationService.authorize(authnRequest).getOrThrow()
                     respondRedirect(result.url)
                 }
 
                 request.url.fullPath.startsWith(tokenEndpointPath) -> {
                     val requestBody = request.body.toByteArray().decodeToString()
-                    val params: TokenRequestParameters = requestBody.decodeFromPostBody()
+                    val params: TokenRequestParameters = requestBody.decodeFromPostBody<TokenRequestParameters>()
                     val result = authorizationService.token(params, request.toRequestInfo()).getOrThrow()
                     respond(
-                        vckJsonSerializer.encodeToString(result),
+                        vckJsonSerializer.encodeToString<TokenResponseParameters>(result),
                         headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                     )
                 }
@@ -333,7 +315,7 @@ class OpenId4VciClientTest : FunSpec() {
                 request.url.fullPath.startsWith(nonceEndpointPath) -> {
                     val result = credentialIssuer.nonce().getOrThrow()
                     respond(
-                        vckJsonSerializer.encodeToString(result),
+                        vckJsonSerializer.encodeToString<ClientNonceResponse>(result),
                         headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                     )
                 }
@@ -344,7 +326,7 @@ class OpenId4VciClientTest : FunSpec() {
                     val params = CredentialRequestParameters.deserialize(requestBody).getOrThrow()
                     val result = credentialIssuer.credential(authn, params, request.toRequestInfo()).getOrThrow()
                     respond(
-                        vckJsonSerializer.encodeToString(result),
+                        vckJsonSerializer.encodeToString<CredentialResponseParameters>(result),
                         headers = headersOf(HttpHeaders.ContentType, ContentType.Application.Json.toString())
                     )
 
@@ -353,7 +335,26 @@ class OpenId4VciClientTest : FunSpec() {
                 else -> respondError(HttpStatusCode.NotFound)
                     .also { Napier.w("NOT MATCHED ${request.url.fullPath}") }
             }
-        }, credentialIssuer, authorizationService)
+        }
+        val clientId = "https://example.com/rp"
+        client = OpenId4VciClient(
+            engine = mockEngine,
+            loadClientAttestationJwt = {
+                BuildClientAttestationJwt(
+                    SignJwt(EphemeralKeyWithSelfSignedCert(), JwsHeaderCertOrJwk()),
+                    clientId = clientId,
+                    issuer = "issuer",
+                    clientKey = clientAuthKeyMaterial.jsonWebKey
+                ).serialize()
+            },
+            signClientAttestationPop = SignJwt(clientAuthKeyMaterial, JwsHeaderNone()),
+            signDpop = SignJwt(dpopKeyMaterial, JwsHeaderJwk()),
+            dpopAlgorithm = dpopKeyMaterial.signatureAlgorithm.toJwsAlgorithm().getOrThrow(),
+            oid4vciService = WalletService(
+                clientId = clientId,
+                keyMaterial = credentialKeyMaterial,
+            ),
+        )
     }
 
     private fun HttpRequestData.toRequestInfo(): RequestInfo = RequestInfo(
@@ -365,19 +366,4 @@ class OpenId4VciClientTest : FunSpec() {
     )
 
     private fun dummyUser(): OidcUserInfoExtended = OidcUserInfoExtended.deserialize("{\"sub\": \"foo\"}").getOrThrow()
-
-    // If the countdownLatch has been unlocked, the correct credential has been stored, and we're done!
-    private suspend fun assertCorrectCredentialIssued() {
-        withContext(Dispatchers.Default.limitedParallelism(1)) {
-            withTimeout(5000) {
-                countdownLatch.lock()
-            }
-        }
-    }
 }
-
-data class SetupResult(
-    val openId4VciClient: OpenId4VciClient,
-    val credentialIssuer: CredentialIssuer,
-    val authorizationService: SimpleAuthorizationService,
-)